Upgrade existing Fleet release with Rancher values
When setting `FleetBeforeRancher` is enabled, Rancher is now able to
detect a pre-existent Fleet release. If such a release is found, Rancher
will apply settings updates to that release, instead of installing the
version pinned to the Rancher release.
That Fleet release can still be upgraded manually through the Rancher
UI, or by disabling the `FleetBeforeRancher` flag.